Added GUI, fixed DCC logic.
GUI includes 4 buttons (left, right, select and e stop) and a two digit
SSD. When a train is selected, it is added to a list and the trains in
that list will have their speed set every update (controlled by the main
loop). An E stop will remove all trains from the list and broadcast the
stop message.